Fixed datatype mismatches
* comparing an unsigned value < 0 is always false
* comparing signed and unsigned expressions can lead to bugs
* using unsigned char to cound up to 255 leads to infinite loops

Found by GCC

+++ libs/dmessage/dmsg.c
@@ -714,7 +714,7 @@ int _dmsg_encrypt_chunk(dmime_message_chunk_t *chunk, dmime_kekset_t *keks) { //
 	if((res = _encrypt_aes_256(outbuf, &(chunk->data[0]), data_size, temp.aes_key, temp.iv)) < 0) {
 		_secure_wipe((unsigned char *)&temp, sizeof(temp));
 		RET_ERROR_INT(ERR_UNSPEC, "error encrypting data");
-	} else if(res != data_size) {
+	} else if((size_t)res != data_size) {
 		RET_ERROR_INT(ERR_UNSPEC, "encrypted an unexpected number of bytes");
 	}
 
@@ -2183,7 +2183,7 @@ dmime_message_chunk_t * _dmsg_decrypt_chunk(dmime_message_chunk_t *chunk, dmime_
 		_secure_wipe(&keyslot_dec, sizeof(dmime_keyslot_t));
 		free(data);
 		RET_ERROR_PTR(ERR_UNSPEC, "an error occurred while decrypting a chunk payload");
-	} else if(res != payload_size) {
+	} else if((size_t)res != payload_size) {
 		_secure_wipe(&keyslot_dec, sizeof(dmime_keyslot_t));
 		free(data);
 		RET_ERROR_PTR(ERR_UNSPEC, "decrypted an unexpected number of bytes");
